requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
openai>=1.3.5
python-slugify>=8.0.1
httpx>=0.23.0,<0.24.0
supabase-py>=0.0.2
websockets>=11.0.3
